+++ internal/broadcast/manager_test.go
@@ -1,0 +1,7 @@
+package broadcast
+
+import "testing"
+
+func TestServerConnections(t *testing.T) {
+
+}